Added additional tests for the new function implementation
